# added softmax!. reduced softmax's memory usage.