--- a/src/softmax.jl
+++ b/src/softmax.jl
@@ -32,16 +32,16 @@ julia> softmax([1, 2, 3])
 
 See also [`logsoftmax`](@ref).
 """
-softmax(x; dims = 1) = softmax!(similar(x), x; dims = dims)
+softmax(x; dims = 1) = softmax!(similar(x, (float ∘ eltype)(x)), x; dims = dims)
 softmax!(x; dims = 1) = softmax!(x, x; dims = dims)
-function softmax!(out::T, x::T; dims = 1) where {T<:AbstractArray}
+function softmax!(out::O, x::T; dims = 1) where {O<:AbstractArray,T<:AbstractArray}
     out .= exp.(x .- maximum(x; dims = dims))
     out ./= sum(out; dims = dims)
 end
 
 ∇softmax(Δ, x; dims = 1) = ∇softmax!(similar(Δ), Δ, x; dims = dims)
 ∇softmax!(Δ, x; dims = 1) = ∇softmax!(Δ, Δ, x; dims = dims)
-function ∇softmax!(out::T, Δ::T, x::T; dims = 1) where {T<:AbstractArray}
+function ∇softmax!(out::O, Δ::O, x::T; dims = 1) where {O<:AbstractArray,T<:AbstractArray}
     softmax!(out, x; dims = dims)
     out .*= Δ .- sum(Δ .* out; dims = dims)
 end
@@ -60,18 +60,23 @@ It is semantically equivalent to the following:
 
 See also [`softmax`](@ref).
 """
-logsoftmax(x; dims = 1) = logsoftmax!(similar(x), x; dims = dims)
+logsoftmax(x; dims = 1) = logsoftmax!(similar(x, (float ∘ eltype)(x)), x; dims = dims)
 logsoftmax!(x; dims = 1) = logsoftmax!(x, x; dims = dims)
-function logsoftmax!(out::T, x::T; dims = 1) where {T<:AbstractArray}
+function logsoftmax!(out::O, x::T; dims = 1) where {O<:AbstractArray,T<:AbstractArray}
     out .= x .- maximum(x; dims = dims)
-    # out .= out .- log.(sum(exp.(out); dims = dims))  # WARN: this will decrease performance.
+    # out .-= log.(sum(exp.(out); dims = dims))  # WARN: this will decrease performance.
     log_ = log.(sum(exp.(out); dims = dims))
     out .-= log_
 end
 
 ∇logsoftmax(Δ, x; dims = 1) = ∇logsoftmax!(similar(Δ), Δ, x; dims = dims)
 ∇logsoftmax!(Δ, x; dims = 1) = ∇logsoftmax!(Δ, Δ, x; dims = dims)
-function ∇logsoftmax!(out::T, Δ::T, x::T; dims = 1) where {T<:AbstractArray}
+function ∇logsoftmax!(
+    out::O,
+    Δ::O,
+    x::T;
+    dims = 1,
+) where {O<:AbstractArray,T<:AbstractArray}
     out .= Δ .- sum(Δ, dims = dims) .* softmax!(out, x; dims = dims)
 end