1. What about add a inplace version of softmax, with `dims` parameter?
2. `softmax` is updated too, this can reduce memory usage by 40%.

note:
The memory uses of `softmax!` is the same as the old `softmax`.
The performance does not change.