@CarloLucibello 2020-12-18
I think both changes are ok. 
Since you are at it, would you mind taking care also of #249?
To make it easier, we could probably replace all those explicit loops with functional/vectorized code, 
although we have to make sure it won't impact performances.