@norci 2020-12-19
I'm glad to help.

I have updated all the function, and refactored the test cases by removing duplicate code.

The performance has increased a little, the usage of memory has reduced a lot.

# benchmark code
```julia
using BenchmarkTools, NNlib
x = rand(1000,1000,16);
Δ = rand(1000,1000,16);
macro mb(expr)
    :(median(@benchmark $expr samples=9 evals=1))
end
@mb softmax(x)
@mb logsoftmax(x)
@mb ∇softmax(Δ, x)
@mb ∇logsoftmax(Δ, x)
```

# benchmark results
## `julia> @mb softmax(x)`
### this
```
  time:             166.028 ms
  gctime:           0.000 ns (0.00%)
  memory:           122.31 MiB
  allocs:           9
```
### original
```
  time:             174.805 ms
  gctime:           3.930 ms (2.25%)
  memory:           244.39 MiB
  allocs:           11
```
## `julia> @mb logsoftmax(x)`
### this
```
  time:             167.900 ms
  gctime:           4.940 ms (2.94%)
  memory:           244.51 MiB
  allocs:           13
```
### original
```
  time:             184.529 ms
  gctime:           3.947 ms (2.14%)
  memory:           244.51 MiB
  allocs:           13
```
## `julia> @mb ∇softmax(Δ, x)`
### this
```
  time:             213.308 ms
  gctime:           7.133 ms (3.34%)
  memory:           244.51 MiB
  allocs:           13
```
### original
```
  time:             228.071 ms
  gctime:           14.638 ms (6.42%)
  memory:           488.65 MiB
  allocs:           17
```
## `julia> @mb ∇logsoftmax(Δ, x)`
### this
```
  time:             185.306 ms
  gctime:           0.000 ns (0.00%)
  memory:           122.44 MiB
  allocs:           11
```
### original
```
  time:             217.055 ms
  gctime:           11.068 ms (5.10%)
  memory:           366.58 MiB
  allocs:           15
```