@CarloLucibello 2020-12-19
very nice. After addressing the comments, it will be convenient to add some tests with integer inputs