@CarloLucibello 2020-12-19
I just noticed https://github.com/FluxML/NNlib.jl/issues/248, I'm not sure how to handle it, but thought it was mentioning here