@CarloLucibello 2020-12-19
better do #248 in a separate PR where we also bump the minor version