@CarloLucibello 2020-12-20
You could add a test for #249 and we should be ready